*API* Change `receiveItemBarCode` operation name
In my opinion it is pointless to mount the service under
`Flow/Services/Custom` and then, on top of that, still name the
operation `Flow/Services/Custom/receiveItemBarCode` (in our case
`Flow_Services_Custom_receiveItemBarCode` as slashes aren't allowed
according to XSD spec).

I therefore decided to stick with the default name which is
`receiveItemBarCode`.