Implement new field ordering in model
With the previous approach the `as_dict()` method always returned an
empty dict when running in live mode (with mocked up sample data it
still worked). It most probably doesn't work because it isn't supported
to extend from a customized class as I did it with
`ComplexModelOrdered`.

Therefore, I now use the old `_type_info` style to define the order.